+++ src/v0_4/bioformats2raw_layout.rs
@@ -14,6 +14,10 @@ pub struct Bioformats2rawLayout {
 
 #[cfg(test)]
 mod tests {
+    use serde_json::Value;
+
+    use crate::v0_4::Ome;
+
     use super::*;
 
     #[test]
@@ -22,7 +26,9 @@ mod tests {
             env!("CARGO_MANIFEST_DIR"),
             "/ome-zarr/0.4/examples/bf2raw/image.json"
         ));
-        let _bioformats2raw: Bioformats2rawLayout = serde_json::from_str(json).unwrap();
+        let ome_metadata: Value = serde_json::from_str(&json).unwrap();
+        let ome_metadata: Ome = serde_json::from_value(ome_metadata.clone()).unwrap();
+        let _bioformats2raw: Bioformats2rawLayout = ome_metadata.bioformats2raw_layout.unwrap();
     }
 
     #[test]
@@ -31,6 +37,8 @@ mod tests {
             env!("CARGO_MANIFEST_DIR"),
             "/ome-zarr/0.4/examples/bf2raw/plate.json"
         ));
-        let _bioformats2raw: Bioformats2rawLayout = serde_json::from_str(json).unwrap();
+        let ome_metadata: Value = serde_json::from_str(&json).unwrap();
+        let ome_metadata: Ome = serde_json::from_value(ome_metadata.clone()).unwrap();
+        let _bioformats2raw: Bioformats2rawLayout = ome_metadata.bioformats2raw_layout.unwrap();
     }
 }