+++ src/v0_5/labels.rs
@@ -23,7 +23,7 @@ pub struct ImageLabel {
 
 #[cfg(test)]
 mod tests {
-    use crate::v0_5::get_ome_attribute_from_zarr_group_metadata;
+    use crate::v0_5::{get_ome_attribute_from_zarr_group_metadata, Ome};
 
     use super::*;
 
@@ -36,7 +36,7 @@ mod tests {
         let group_metadata: serde_json::Map<String, serde_json::Value> =
             serde_json::from_str(json).unwrap();
         let ome_metadata = get_ome_attribute_from_zarr_group_metadata(&group_metadata).unwrap();
-        let image_label = ome_metadata.get("image-label").unwrap();
-        let _image_label: ImageLabel = serde_json::from_value(image_label.clone()).unwrap();
+        let ome_metadata: Ome = serde_json::from_value(ome_metadata.clone()).unwrap();
+        let _image_label: ImageLabel = ome_metadata.image_label.unwrap();
     }
 }