+++ src/v0_5/plate.rs
@@ -31,7 +31,7 @@ pub struct Plate {
 
 #[cfg(test)]
 mod tests {
-    use crate::v0_5::get_ome_attribute_from_zarr_group_metadata;
+    use crate::v0_5::{get_ome_attribute_from_zarr_group_metadata, Ome};
 
     use super::*;
 
@@ -44,8 +44,8 @@ mod tests {
         let group_metadata: serde_json::Map<String, serde_json::Value> =
             serde_json::from_str(json).unwrap();
         let ome_metadata = get_ome_attribute_from_zarr_group_metadata(&group_metadata).unwrap();
-        let plate = ome_metadata.get("plate").unwrap();
-        let _plate: Plate = serde_json::from_value(plate.clone()).unwrap();
+        let ome_metadata: Ome = serde_json::from_value(ome_metadata.clone()).unwrap();
+        let _plate: Plate = ome_metadata.plate.unwrap();
     }
 
     #[test]
@@ -57,7 +57,7 @@ mod tests {
         let group_metadata: serde_json::Map<String, serde_json::Value> =
             serde_json::from_str(json).unwrap();
         let ome_metadata = get_ome_attribute_from_zarr_group_metadata(&group_metadata).unwrap();
-        let plate = ome_metadata.get("plate").unwrap();
-        let _plate: Plate = serde_json::from_value(plate.clone()).unwrap();
+        let ome_metadata: Ome = serde_json::from_value(ome_metadata.clone()).unwrap();
+        let _plate: Plate = ome_metadata.plate.unwrap();
     }
 }