+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.2] - 2024-12-10
+
 ### Added
 - Add top-level OME-NGFF/OME-Zarr structs
   - Add `v0_4::OmeNgffGroupAttributes`
@@ -32,7 +34,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial public release
 
-[unreleased]: https://github.com/LDeakin/rust_ome_zarr_metadata/compare/v0.2.1...HEAD
+[unreleased]: https://github.com/LDeakin/rust_ome_zarr_metadata/compare/v0.2.2...HEAD
+[0.2.2]: https://github.com/LDeakin/rust_ome_zarr_metadata/releases/tag/v0.2.2
 [0.2.1]: https://github.com/LDeakin/rust_ome_zarr_metadata/releases/tag/v0.2.1
 [0.2.0]: https://github.com/LDeakin/rust_ome_zarr_metadata/releases/tag/v0.2.0
 [0.1.0]: https://github.com/LDeakin/rust_ome_zarr_metadata/releases/tag/v0.1.0